In order for one format specifier to work for both architectures, you must use a long specifier and cast the value to long.

+++ ILSC-Brisbane/PropertyListReader.m
@@ -23,7 +23,7 @@
                                                 format:&format
                                                 error:&errorDesc];
     if (!properties) {
-        NSLog(@"Error reading plist: %@, format: %u", errorDesc, format);
+        NSLog(@"Error reading plist: %@, format: %lu", errorDesc, (unsigned long)format);
     }
 
     return properties;
